+++ static/sql/sqlite3/schema.sql
