+++ test.sh
@@ -2,14 +2,18 @@
 
 set -e
 
-trap 'kill %1; kill %2' ERR EXIT
+trap 'kill %1; kill %2; kill %3' ERR EXIT
 
-./transfer2go -config test/config/config1.json -auth=false >/dev/null 2>&1 &
-
-sleep 1
-
-./transfer2go -config test/config/config2.json -auth=false -agent http://localhost:8989 >/dev/null 2>&1 &
+./transfer2go -config test/config/main.json -auth=false >/dev/null 2>&1 &
 
 sleep 1
 
-cd test && go test request_test.go
+./transfer2go -config test/config/source.json -auth=false -agent http://localhost:8989 >/dev/null 2>&1 &
+
+sleep 1
+
+./transfer2go -config test/config/destination.json -auth=false -agent http://localhost:8989 >/dev/null 2>&1 &
+
+sleep 1
+
+cd test && go test function_test.go
