--- test/catalog/catalog1.json
@@ -1,4 +1,0 @@
-{
-    "type":"sqlite3",
-    "uri":"test/catalog/catalog.db"
-}
