--- test/catalog/catalog2.json
@@ -1,5 +1,0 @@
-{
-    "type":"sqlite3",
-    "uri":"test/catalog/catalog2.db"
-}
-
