+++ test/catalog/main.json
@@ -1,0 +1,4 @@
+{
+    "type":"sqlite3",
+    "uri":"test/catalog/main.db"
+}
