+++ test/config/destination.json
@@ -1,14 +1,14 @@
 {
-    "catalog":"test/catalog/catalog2.json",
+    "catalog":"test/catalog/destination.json",
     "protocol":"cp",
     "backend":"test/",
     "tool":"/bin/cp",
-    "url":"http://localhost:8000",
-    "port": 8000,
+    "url":"http://localhost:9000",
+    "port": 9000,
     "mfile":"metrics2.log",
     "minterval":60,
     "verbose":0,
-    "name":"Test2",
+    "name":"destination",
     "staticdir":"static",
     "type": "push"
 }
