+++ test/config/source.json
@@ -1,0 +1,14 @@
+{
+    "catalog":"test/catalog/source.json",
+    "protocol":"cp",
+    "backend":"test/",
+    "tool":"/bin/cp",
+    "url":"http://localhost:8000",
+    "port": 8000,
+    "mfile":"metrics2.log",
+    "minterval":60,
+    "verbose":0,
+    "name":"source",
+    "staticdir":"static",
+    "type": "push"
+}
